reorganize gui file structure
