+++ Resources/gui_tags.py
@@ -8,3 +8,4 @@ REGEX_WIZARD_MODAL_TAG = "modal_regex_wizard"
 REGEX_WIZARD_MODAL_COMBO_TYPE_TAG = "modal_combo_item_type"
 REGEX_WIZARD_MODAL_COMBO_BASE_TAG = "modal_combo_item_base"
 REGEX_WIZARD_MODAL_COMBO_REGEX_TAG = "modal_combo_regex"
+CRAFTING_TARGET_COMBO_TAG = "crafting_target_selector_tag"
