+++ src/gui_layout.ini
@@ -48,14 +48,14 @@ Collapsed=0
 DockId=0x0000000D,0
 
 [Window][###30]
-Pos=0,313
-Size=1264,311
+Pos=0,365
+Size=1264,328
 Collapsed=0
 DockId=0x00000011,0
 
 [Window][###29]
 Pos=0,142
-Size=1264,169
+Size=1264,221
 Collapsed=0
 DockId=0x0000000F,0
 
@@ -67,7 +67,7 @@ DockId=0x00000013,0
 
 [Window][###40]
 Pos=489,270
-Size=430,265
+Size=300,450
 Collapsed=0
 
 [Docking][Data]
@@ -86,10 +86,10 @@ DockSpace                     ID=0x7C6B3D9B Window=0xA87D555D Pos=0,0 Size=1264,
             DockNode          ID=0x0000000C Parent=0x0000000A SizeRef=784,328 Split=Y
               DockNode        ID=0x0000000D Parent=0x0000000C SizeRef=784,140 Selected=0xA4B861D9
               DockNode        ID=0x0000000E Parent=0x0000000C SizeRef=784,469 Split=Y
-                DockNode      ID=0x0000000F Parent=0x0000000E SizeRef=784,169 Selected=0x99D84869
-                DockNode      ID=0x00000010 Parent=0x0000000E SizeRef=784,448 Split=Y
-                  DockNode    ID=0x00000011 Parent=0x00000010 SizeRef=784,311 Selected=0x5F94F9BD
-                  DockNode    ID=0x00000012 Parent=0x00000010 SizeRef=784,135 Split=Y
+                DockNode      ID=0x0000000F Parent=0x0000000E SizeRef=784,221 Selected=0x99D84869
+                DockNode      ID=0x00000010 Parent=0x0000000E SizeRef=784,396 Split=Y
+                  DockNode    ID=0x00000011 Parent=0x00000010 SizeRef=784,328 Selected=0x5F94F9BD
+                  DockNode    ID=0x00000012 Parent=0x00000010 SizeRef=784,66 Split=Y
                     DockNode  ID=0x00000013 Parent=0x00000012 SizeRef=784,294 Selected=0x62F4D00D
                     DockNode  ID=0x00000014 Parent=0x00000012 SizeRef=784,32 CentralNode=1
